+++ src/math/powf.c
@@ -219,7 +219,7 @@ float powf(float x, float y)
 			return sn*huge*huge;  /* overflow */
 	} else if ((j&0x7fffffff) > 0x43160000)  /* z < -150 */ // FIXME: check should be  (uint32_t)j > 0xc3160000
 		return sn*tiny*tiny;  /* underflow */
-	else if (j == 0xc3160000) {  /* z == -150 */
+	else if (j == (int32_t)0xc3160000) {  /* z == -150 */
 		if (p_l <= z-p_h)
 			return sn*tiny*tiny;  /* underflow */
 	}
